10-agentic-codingv2/back-end/requirements.txt
# Core Dependencies
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-prebuilt==1.0.2
langchainhub>=0.1.0

# OpenAI
openai==2.7.2

# Tools and Utilities
python-dotenv==1.2.1
wikipedia==1.4.0
arxiv==2.3.1

# Vector Store
faiss-cpu==1.12.0

# Web Scraping
beautifulsoup4==4.14.2

# HTTP Client
httpx==0.28.1

# Google Serper API (for search functionality)
google-search-results>=2.4.0

# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.4
pydantic-settings==2.12.0

# Rate Limiting
slowapi==0.1.9

# Console Output
colorama==0.4.6

# Additional HTTP Client for health checks
requests==2.32.5
